+++ nvdaHelper/remote/sconscript
@@ -95,7 +95,7 @@ remoteLib=env.SharedLibrary(
 		controllerInternalRPCClientSource,
 		"sysListView32.cpp",
 		"winword.cpp",
-		"WinWord/Links.cpp",
+		"WinWord/Fields.cpp",
 		"gdiHooks.cpp",
 		"displayModel.cpp",
 		"displayModelRemote.cpp",
